Renamed CommandCog to ClientCog, migrated application command decorators to Client, removed CommandClient, moved everything remaining from nextcord.command_client to nextcord.application_command
