+++ nextcord/__init__.py
@@ -60,7 +60,7 @@ from .interactions import *
 from .components import *
 from .threads import *
 from .health_check import *
-from .command_client import *
+from .application_command import *
 
 
 class VersionInfo(NamedTuple):
