+++ nextcord/ext/commands/bot.py
@@ -36,7 +36,6 @@ import types
 from typing import Any, Callable, Mapping, List, Dict, TYPE_CHECKING, Optional, TypeVar, Type, Union
 
 import nextcord
-import nextcord.command_client
 
 from .core import GroupMixin
 from .view import StringView
@@ -556,7 +555,7 @@ class BotBase(GroupMixin):
 
         cog = cog._inject(self)
         self.__cogs[cog_name] = cog
-        # TODO: Mirror this in CommandClient so it actually lines up. Also put it somewhere that actually makes sense.
+        # TODO: This blind call to nextcord.Client is dumb.
         super().add_cog(cog)
 
     def get_cog(self, name: str) -> Optional[Cog]:
@@ -1038,8 +1037,8 @@ class BotBase(GroupMixin):
     async def on_message(self, message):
         await self.process_commands(message)
 
-# class Bot(BotBase, nextcord.Client):
-class Bot(BotBase, nextcord.command_client.CommandClient):
+
+class Bot(BotBase, nextcord.Client):
     """Represents a discord bot.
 
     This class is a subclass of :class:`nextcord.Client` and as a result
@@ -1110,6 +1109,7 @@ class Bot(BotBase, nextcord.command_client.CommandClient):
         .. versionadded:: 1.7
     """
     pass
+
 
 class AutoShardedBot(BotBase, nextcord.AutoShardedClient):
     """This is similar to :class:`.Bot` except that it is inherited from
