+++ nextcord/state.py
@@ -75,8 +75,6 @@ if TYPE_CHECKING:
     from .types.sticker import GuildSticker as GuildStickerPayload
     from .types.guild import Guild as GuildPayload
     from .types.message import Message as MessagePayload
-    from .application_command import ApplicationCommandResponse
-    from .command_client import ApplicationCommand
 
     T = TypeVar('T')
     CS = TypeVar('CS', bound='ConnectionState')
@@ -307,36 +305,6 @@ class ConnectionState:
         ret = Intents.none()
         ret.value = self._intents.value
         return ret
-
-    @property
-    def application_commands(self) -> List[ApplicationCommandResponse]:
-        return list(self._application_commands.values())
-
-    # def _add_application_command(self, app_cmd: ApplicationCommandResponse):
-    #     self._application_commands[app_cmd.id] = app_cmd
-    #     if app_cmd.guild_id:
-    #         if app_cmd.guild_id not in self._guild_application_command_names:
-    #             self._guild_application_command_names[app_cmd.guild_id] = dict()
-    #         self._guild_application_command_names[app_cmd.guild_id][app_cmd.name] = app_cmd
-    #     else:
-    #         self._global_applicaiton_command_names[app_cmd.name] = app_cmd
-    #
-    # def get_application_command(self, cmd_id: int) -> Optional[ApplicationCommandResponse]:
-    #     return self._application_commands.get(cmd_id, None)
-    #
-    # def get_guild_application_command(self, guild_id: int, name: str) -> Optional[ApplicationCommandResponse]:
-    #     return self._guild_application_command_names.get(guild_id, dict()).get(name, None)
-    #
-    # def get_global_application_command(self, name: str) -> Optional[ApplicationCommandResponse]:
-    #     return self._global_applicaiton_command_names.get(name, None)
-    #
-    # def _remove_application_command(self, cmd_id: int):
-    #     app_cmd = self._application_commands.pop(cmd_id, None)
-    #     if app_cmd:
-    #         if app_cmd.guild_id:
-    #             self._guild_application_command_names[app_cmd.guild_id].pop(app_cmd.name, None)
-    #         else:
-    #             self._global_applicaiton_command_names.pop(app_cmd.name, None)
 
     @property
     def voice_clients(self) -> List[VoiceProtocol]:
